Isolate IAM delegation DDL actors
