+++ ydb/core/kqp/gateway/behaviour/external_data_source/iam_delegation_ddl.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "manager.h"
+
+namespace NKikimr::NKqp::NExternalDataSource {
+
+bool IsIamDelegationEnabled(NActors::TActorSystem* actorSystem);
+
+TExternalDataSourceManager::TYqlConclusionStatus PrepareIamDelegation(
+    NKikimrSchemeOp::TExternalDataSourceDescription& description,
+    TStringBuf name);
+
+TExternalDataSourceManager::TAsyncStatus ExecuteIamDelegationDdl(
+    const NKikimrSchemeOp::TModifyScheme& schemeTx,
+    const TExternalDataSourceManager::TExternalModificationContext& context,
+    NKqpProto::TKqpSchemeOperation::OperationCase operationCase);
+
+} // namespace NKikimr::NKqp::NExternalDataSource
